# frontend/static/openapi.json
{
	"openapi": "3.1.1",
	"info": {
		"title": "Kadia explorer workflow API",
		"version": "1.0.0",
		"description": "Versioned documentation for the 23 operations supported by the portable SDK. Other server routes are outside this contract subset. Amounts and difficulty remain decimal strings; heights/counts are numbers. Exact historical address queries require full retained history; network history requires the entire requested range to be retained. Contract version is distinct from chain height and software release. Compatible fields may be added; unknown enum values should be surfaced rather than guessed."
	},
	"servers": [
		{
			"url": "/v1"
		}
	],
	"paths": {
...
		"/mempool": {
			"get": {
				"operationId": "mempool",
				"summary": "mempool",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Observation. Read its source, completeness, timestamps and snapshot fields. Mempool entries are pending on one configured node; they are not indexed confirmation, ownership or network-wide acceptance claims.",
						"headers": {
							"x-explorer-completeness": {
								"schema": {
									"type": "string",
									"enum": ["complete", "incomplete"]
								},
								"description": "Coverage of the store read. Omitted for source-only/status responses."
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MempoolSnapshot"
								}
							}
						}
					},
					"default": {
						"description": "HTTP failures include invalid input (400), missing data (404), timeout (408), changed snapshot (409), body limit (413), read budget (422), throttling (429), unavailable history/source (503), or integrity failure (500). Framework errors may have no JSON body. The SDK preserves HTTP status and does not automatically retry.",
						"content": {
							"application/problem+json": {
								"schema": {
									"type": "object",
									"properties": {
										"type": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"status": {
											"type": "integer"
										},
										"detail": {
											"type": "string"
										},
										"code": {
											"type": "string"
										}
									},
									"required": ["title", "status", "detail"]
								}
							}
						}
					}
				}
			}
		},
...
		"/txs/{id}": {
			"get": {
				"operationId": "transaction",
				"summary": "transaction",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"pattern": "^[a-fA-F0-9]{64}$"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Observation. Read its source, completeness, timestamps and snapshot fields. Mempool entries are pending on one configured node; they are not indexed confirmation, ownership or network-wide acceptance claims.",
						"headers": {
							"x-explorer-completeness": {
								"schema": {
									"type": "string",
									"enum": ["complete", "incomplete"]
								},
								"description": "Coverage of the store read. Omitted for source-only/status responses."
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TxDto"
								}
							}
						}
					},
					"default": {
						"description": "HTTP failures include invalid input (400), missing data (404), timeout (408), changed snapshot (409), body limit (413), read budget (422), throttling (429), unavailable history/source (503), or integrity failure (500). Framework errors may have no JSON body. The SDK preserves HTTP status and does not automatically retry.",
						"content": {
							"application/problem+json": {
								"schema": {
									"type": "object",
									"properties": {
										"type": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"status": {
											"type": "integer"
										},
										"detail": {
											"type": "string"
										},
										"code": {
											"type": "string"
										}
									},
									"required": ["title", "status", "detail"]
								}
							}
						}
					}
				}
			}
		},
...
		"/tx-summaries": {
			"get": {
				"operationId": "transactions",
				"summary": "transactions",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 500
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"maxLength": 4096
						}
					},
					{
						"name": "snapshot",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"maxLength": 4096
						}
					},
					{
						"name": "dir",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["asc", "desc"]
						}
					},
					{
						"name": "consistency",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["strict", "best_effort"],
							"default": "best_effort"
						},
						"description": "The SDK always requests strict consistency and carries cursor and snapshot together."
					}
				],
				"responses": {
					"200": {
						"description": "Observation. Read its source, completeness, timestamps and snapshot fields. Mempool entries are pending on one configured node; they are not indexed confirmation, ownership or network-wide acceptance claims.",
						"headers": {
							"x-explorer-completeness": {
								"schema": {
									"type": "string",
									"enum": ["complete", "incomplete"]
								},
								"description": "Coverage of the store read. Omitted for source-only/status responses."
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"items": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/TxSummaryDto"
											}
										},
										"next_cursor": {
											"anyOf": [
												{
													"type": "string"
												},
												{
													"type": "null"
												}
											]
										},
										"next_snapshot": {
											"anyOf": [
												{
													"type": "string"
												},
												{
													"type": "null"
												}
											]
										},
										"consistency": {
											"anyOf": [
												{
													"type": "string",
													"const": "strict"
												},
												{
													"type": "string",
													"const": "best_effort"
												}
											]
										},
										"anchor": {
											"anyOf": [
												{
													"type": "object",
													"properties": {
														"height": {
															"type": "number"
														},
														"block_id": {
															"type": "string"
														}
													},
													"required": ["height", "block_id"]
												},
												{
													"type": "null"
												}
											]
										},
										"observed_anchor": {
											"anyOf": [
												{
													"type": "object",
													"properties": {
														"height": {
															"type": "number"
														},
														"block_id": {
															"type": "string"
														}
													},
													"required": ["height", "block_id"]
												},
												{
													"type": "null"
												}
											]
										}
									},
									"required": ["items", "next_cursor"]
								}
							}
						}
					},
					"default": {
						"description": "HTTP failures include invalid input (400), missing data (404), timeout (408), changed snapshot (409), body limit (413), read budget (422), throttling (429), unavailable history/source (503), or integrity failure (500). Framework errors may have no JSON body. The SDK preserves HTTP status and does not automatically retry.",
						"content": {
							"application/problem+json": {
								"schema": {
									"type": "object",
									"properties": {
										"type": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"status": {
											"type": "integer"
										},
										"detail": {
											"type": "string"
										},
										"code": {
											"type": "string"
										}
									},
									"required": ["title", "status", "detail"]
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"MempoolSnapshot": {
				"type": "object",
				"properties": {
					"scope": {
						"type": "string",
						"const": "configured_node_mempool"
					},
					"source": {
						"type": "string",
						"const": "configured_primary_node"
					},
					"checked_at_ms": {
						"type": "number"
					},
					"expires_at_ms": {
						"type": "number"
					},
					"cached": {
						"type": "boolean"
					},
					"limit": {
						"type": "number",
						"const": 100
					},
					"limit_reached": {
						"type": "boolean"
					},
					"observed_count": {
						"type": "number"
					},
					"items": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/PendingTransaction"
						}
					}
				},
				"required": [
					"scope",
					"source",
					"checked_at_ms",
					"expires_at_ms",
					"cached",
					"limit",
					"limit_reached",
					"observed_count",
					"items"
				]
			},
			"PendingTransaction": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"input_count": {
						"type": "number"
					},
					"data_input_count": {
						"type": "number"
					},
					"output_count": {
						"type": "number"
					},
					"size": {
						"anyOf": [
							{
								"type": "number"
							},
							{
								"type": "null"
							}
						]
					},
					"fee": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"required": ["id", "input_count", "data_input_count", "output_count", "size", "fee"]
			},
...
			"TxDto": {
				"type": "object",
				"properties": {
					"block_id": {
						"type": "string"
					},
					"indexed_height": {
						"anyOf": [
							{
								"type": "number"
							},
							{
								"type": "null"
							}
						]
					},
					"confirmations": {
						"anyOf": [
							{
								"type": "number"
							},
							{
								"type": "null"
							}
						]
					},
					"id": {
						"type": "string"
					},
					"height": {
						"type": "number"
					},
					"index": {
						"type": "number"
					},
					"timestamp": {
						"type": "number"
					},
					"size": {
						"type": "number"
					},
					"fee": {
						"type": "string"
					},
					"inputs": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/InputDto"
						}
					},
					"data_inputs": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"outputs": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/BoxDto"
						}
					}
				},
				"required": [
					"id",
					"height",
					"index",
					"timestamp",
					"size",
					"fee",
					"inputs",
					"data_inputs",
					"outputs"
				]
			},
			"InputDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"box": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/BoxDto"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"required": ["id", "box"]
			},
			"BoxDto": {
				"type": "object",
				"properties": {
					"kind": {
						"anyOf": [
							{
								"type": "string",
								"const": "box"
							},
							{
								"type": "string",
								"const": "fee"
							},
							{
								"type": "string",
								"const": "emission"
							}
						]
					},
					"id": {
						"type": "string"
					},
					"tx_id": {
						"type": "string"
					},
					"index": {
						"type": "number"
					},
					"value": {
						"type": "string"
					},
					"creation_height": {
						"type": "number"
					},
					"ergo_tree": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						]
					},
					"address": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						]
					},
					"template_hash": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						]
					},
					"tree_hash": {
						"type": "string"
					},
					"tokens": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/TokenDto"
						}
					},
					"registers": {
						"anyOf": [
							{
								"type": "object",
								"additionalProperties": {}
							},
							{
								"type": "null"
							}
						]
					},
					"size": {
						"type": "number"
					},
					"spent_by": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						]
					},
					"spent_height": {
						"anyOf": [
							{
								"type": "number"
							},
							{
								"type": "null"
							}
						]
					},
					"rent": {
						"$ref": "#/components/schemas/RentDto"
					}
				},
				"required": [
					"id",
					"tx_id",
					"index",
					"value",
					"creation_height",
					"ergo_tree",
					"address",
					"template_hash",
					"tree_hash",
					"tokens",
					"registers",
					"size",
					"spent_by",
					"spent_height",
					"rent"
				]
			},
			"TokenDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"amount": {
						"type": "string"
					},
					"name": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						]
					},
					"decimals": {
						"anyOf": [
							{
								"type": "number"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"required": ["id", "amount", "name", "decimals"]
			},
			"RentDto": {
				"type": "object",
				"properties": {
					"maturity_height": {
						"type": "number"
					},
					"due_nano": {
						"type": "string"
					},
					"claimable_at_tip": {
						"type": "boolean"
					},
					"consensus_fee_nano": {
						"type": "string"
					},
					"collectible": {
						"type": "boolean"
					}
				},
				"required": [
					"maturity_height",
					"due_nano",
					"claimable_at_tip",
					"consensus_fee_nano",
					"collectible"
				]
			},
...
			"TxSummaryDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"height": {
						"type": "number"
					},
					"index": {
						"type": "number"
					},
					"timestamp": {
						"type": "number"
					},
					"size": {
						"type": "number"
					},
					"fee": {
						"type": "string"
					},
					"input_count": {
						"type": "number"
					},
					"data_input_count": {
						"type": "number"
					},
					"output_count": {
						"type": "number"
					}
				},
				"required": [
					"id",
					"height",
					"index",
					"timestamp",
					"size",
					"fee",
					"input_count",
					"data_input_count",
					"output_count"
				]
			},
...
		}
	}
}